Feat: 코스 추천 구현

+++ src/main/java/com/example/ondongnae/backend/course/dto/RecommendedCourseStoreDto.java
@@ -9,6 +9,6 @@ public class RecommendedCourseStoreDto {
 
     private String name;
     private String longDescription;
-    private int order;
+    private long order;
 
 }
